change namespace from sample to app

+++ src/Service/Population.php
@@ -1,20 +1,32 @@
 <?php
 
-namespace Sample\Service;
+namespace App\Service;
 
 use PSX\Http\Exception as StatusCode;
-use Sample\Model\Collection;
-use Sample\Table\Population as TablePopulation;
+use App\Model\Collection;
+use App\Model\Population as PopulationModel;
+use App\Table\Population as TablePopulation;
 
 class Population
 {
+    /**
+     * @var \App\Table\Population
+     */
     protected $populationTable;
 
+    /**
+     * @param \App\Table\Population $populationTable
+     */
     public function __construct(TablePopulation $populationTable)
     {
         $this->populationTable = $populationTable;
     }
 
+    /**
+     * @param int $startIndex
+     * @param int $count
+     * @return \App\Model\Collection
+     */
     public function getAll($startIndex = 0, $count = 16)
     {
         return new Collection(
@@ -23,6 +35,10 @@ class Population
         );
     }
 
+    /**
+     * @param int $id
+     * @return \PSX\Record\Record
+     */
     public function get($id)
     {
         $population = $this->populationTable->get($id);
@@ -34,35 +50,53 @@ class Population
         return $population;
     }
 
-    public function create($place, $region, $count, $users, $worldUsers)
+    /**
+     * @param \App\Model\Population $model
+     */
+    public function create(PopulationModel $model)
     {
         $this->populationTable->create([
-            'place'      => $place,
-            'region'     => $region,
-            'population' => $count,
-            'users'      => $users,
-            'worldUsers' => $worldUsers,
+            'place'      => $model->getPlace(),
+            'region'     => $model->getRegion(),
+            'population' => $model->getPopulation(),
+            'users'      => $model->getUsers(),
+            'worldUsers' => $model->getWorldUsers(),
             'datetime'   => new \DateTime(),
         ]);
     }
 
-    public function update($id, $place, $region, $count, $users, $worldUsers)
+    /**
+     * @param int $id
+     * @param \App\Model\Population $model
+     */
+    public function update($id, PopulationModel $model)
     {
         $population = $this->get($id);
+
+        if (empty($population)) {
+            throw new StatusCode\NotFoundException('Internet population not found');
+        }
 
         $this->populationTable->update([
             'id'         => $population['id'],
-            'place'      => $place,
-            'region'     => $region,
-            'population' => $count,
-            'users'      => $users,
-            'worldUsers' => $worldUsers,
+            'place'      => $model->getPlace(),
+            'region'     => $model->getRegion(),
+            'population' => $model->getPopulation(),
+            'users'      => $model->getUsers(),
+            'worldUsers' => $model->getWorldUsers(),
         ]);
     }
 
+    /**
+     * @param int $id
+     */
     public function delete($id)
     {
         $population = $this->get($id);
+
+        if (empty($population)) {
+            throw new StatusCode\NotFoundException('Internet population not found');
+        }
 
         $this->populationTable->delete([
             'id' => $population['id']
